requirements.txt
allure-pytest==2.8.40
allure-python-commons==2.8.40
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
docopt==0.6.2
idna==2.10
importlib-metadata==3.10.0
iniconfig==1.1.1
jmespath==0.10.0
logzero==1.7.0
packaging==20.9
pipreqs==0.4.10
pluggy==0.13.1
py==1.10.0
PyMySQL==1.0.2
pyparsing==2.4.7
pytest==6.2.3
pytest-assert-utils==0.2.2
pytest-assume==2.4.2
pytest-check==1.0.1
pytest-ordering==0.6
pytz==2021.1
requests==2.25.1
six==1.15.0
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.4
xlrd==2.0.1
xlwt==1.3.0
yarg==0.1.9
zipp==3.4.1
